modifies pytest to use xdist to run in parallel, uses loadfile, serializes tests that have race conditions, and renames tests to avoid namespace collisions
